Move custom css selectors
