+++ src/blocks/button-maxi/custom-css.js
@@ -115,13 +115,3 @@ export const categoriesButton = [
 	'after button',
 	'icon',
 ];
-
-export const selectorsMap = {};
-
-export const categoriesMap = [
-	'map',
-	'before map',
-	'after map',
-	'title',
-	'address',
-];
