+++ src/blocks/map-maxi/custom-css.js
@@ -1,0 +1,7 @@
+export const categoriesMap = [
+	'map',
+	'before map',
+	'after map',
+	'title',
+	'address',
+];
