+++ src/extensions/index.js
@@ -10,4 +10,3 @@ import './text';
 import './style-cards';
 import './column-templates';
 import './svg';
-import './custom-css';
